revert OID typedef to u_long as agreed to on -coders and in the admin meeting
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@20448 06827809-a52a-0410-b366-d66718629ded

+++ include/net-snmp/library/oid.h
@@ -6,7 +6,7 @@
 #endif
 
 #ifndef EIGHTBIT_SUBIDS
-typedef uint32_t oid;
+typedef u_long oid;
 #define MAX_SUBID   0xFFFFFFFF
 #define NETSNMP_PRIo NETSNMP_PRI32
 #else
